@stateful decorator
A handy decorator for automatically saving the state of an object to a
file based on the initialization arguments. The next time the object is
initialized with the same arguments, the state is loaded from the file
automatically, thus saving initialization time. For example, with the
following code:

```python
class LongLoadingClass(object):
    def __init__(self, arg1, arg2):
        # Long loading code

a = LongLoadingClass(123, 456)
b = LongLoadingClass(123, 456)
```

The first instantiation of the class will actually do the
initialization, but the second will load the state previously saved in
the first initialization.

This decorator is useful to avoid repeating the same initialization. For
example, this is very useful in Machine Learning where some
preprocessing doesn't change and thus can be saved once and repeatedly
used.