[ModelTrainer] Optimizer updater + [Corpus] min_word_frequency
Added an option to pass in an optimizer updater to the model trainer.

Added `min_word_frequency` to the `Corpus` class. Also, change the
initializer such that it accepts a line iterable instead of a filepath.
This is important to allow using a normal file, a gzip file, a string
collection, and so on.