[GeneratorWithLen] Accept a function that returns a generator
Instead of just accepting a generator, the class can now also accept a
function that returns a generator when called. One usage of this is to
be able to iterate a function that uses the `yield` keyword more than
once (I mean iterate more than once) by passing the function itself to
this class.