NamedIteratorSlice class (plus named_islice method)
This class builds an iterator that has a custom name returned when str()
or repr() is called on the instance. This is useful when trying to
generate a unique signature that identifies a set of arguments through
the use of the serialization.object_sig() method. Simply using Python`s
islice() method would generate an output like `<itertools.islice object
at 0x7fc20a5473b0>` when calling the str() or repr() methods.

The benefit of such an iterator is when trying to save the state of an
object based on a certain set of arguments that identifies it, one of
them being an iterator. For example, if you are doing some expensive
operations on a large file which you cannot load into memory, you are
likely to use an iterable of some sort to process the files in chunks.
To avoido repeating the same expensive operations, you might want to
cache the result somewhere. To uniquely generate a key for caching, it
is necessary to be able to uniquely identify the iterable, for which
this class can be handy

As for the `named_islice`, it is an islice-like method to return
instances of NamedIteratorSlice .