Disabling stateful-ness in Corpus class
Disabling stateful-ness in Corpus class at this point, since the lines
argument, which is an iterator, might have different values even for
the same content, hence resulting in a new state ID every time. In fact,
even if it is not an iterator, e.g. list, we probably don't want to use its
string representation as part of the ID. We should probably allow the
user to specify an ID or modify the @stateful decorator to only use the
ID as the key of the object.